corrected bars

+++ MeroFaseMar/Assets/Acreano/ScriptsA/BarraNivel.cs
@@ -5,8 +5,8 @@ using UnityEngine.SceneManagement;
 
 public class BarraNivel : MonoBehaviour
 {
-    private int teste = 0;
-    public int qntd;
+    private float teste = 0;
+    public float qntd;
     Scene scene;
     public GameObject gameMgr;
     private GameManager gameManager;
